Split a Config model of Batch
In practice, the `Config` file is a different model to `Batch` and
should be split. This commit delegates the `Batch` methods to a new
Config model
