+++ src/models/batch.py
@@ -1,10 +1,9 @@
 
 import datetime
-import os
 from sqlalchemy import Column, String, Integer, DateTime
-import yaml
 
 from database import Base
+from models.config import Config
 
 class Batch(Base):
     __tablename__ = 'batches'
@@ -14,23 +13,16 @@ class Batch(Base):
     config = Column(String)
     created_date = Column(DateTime, default=datetime.datetime.utcnow)
 
+
     def __name__(self):
-        return os.path.basename(os.path.dirname(self.config))
+        return self.config_model.__name__()
 
     def help(self):
-        default = 'MISSING: Help for {}'.format(self.__name__())
-        self.data.setdefault('help', default)
-        return self.data['help']
+        return self.config_model.help()
 
     def command(self):
-        n = self.__name__()
-        default = 'echo "No command given for: {}"'.format(n)
-        self.data.setdefault('command', default)
-        return self.data['command']
+        return self.config_model.command()
 
     def __init__(self, **kwargs):
-        self.config = kwargs["config"]
-        def __read_data():
-            with open(self.config, 'r') as stream:
-                return yaml.load(stream) or {}
-        self.data = __read_data()
+        self.config = kwargs['config']
+        self.config_model = Config(self.config)
